Revise bulk discounts table to use integer instead of float for discount percent.

+++ db/migrate/20201108014450_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
       t.string :description
-      t.float :discount_percent
+      t.integer :discount_percent
       t.integer :minimum_quantity
       t.references :merchant, foreign_key: true
